refactor: moved main and landing components to pages
