--- frontend/src/components/layout/landing/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Landing';
