--- frontend/src/components/layout/main/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Main';
