+++ frontend/src/pages/landing/LandingPage.css
@@ -12,12 +12,8 @@
   flex-direction: column;
   min-height: 0;
   background:
-    repeating-linear-gradient(0deg,
-      rgb(255 255 255 / 0.03) 0 1px,
-      transparent 1px 48px),
-    repeating-linear-gradient(90deg,
-      rgb(255 255 255 / 0.03) 0 1px,
-      transparent 1px 48px),
+    repeating-linear-gradient(0deg, rgb(255 255 255 / 0.03) 0 1px, transparent 1px 48px),
+    repeating-linear-gradient(90deg, rgb(255 255 255 / 0.03) 0 1px, transparent 1px 48px),
     linear-gradient(180deg, #060912 0%, #070a16 100%);
 }
 
@@ -35,7 +31,7 @@
   background-image:
     linear-gradient(105deg, rgb(6 9 20 / 90%) 0%, rgb(6 9 20 / 48%) 40%, rgb(10 12 24 / 22%) 62%, transparent 82%),
     linear-gradient(180deg, rgb(87 57 230 / 14%) 0%, rgb(10 12 24 / 76%) 100%),
-    url('../../../assets/images/background.png');
+    url('../../assets/images/background.png');
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
@@ -106,22 +102,18 @@
   max-width: 300px;
   height: auto;
   object-fit: contain;
-  filter:
-    drop-shadow(0 0 10px rgb(255 186 79 / 40%)) drop-shadow(0 0 20px rgb(255 140 54 / 26%)) drop-shadow(0 0 36px rgb(142 88 255 / 22%));
+  filter: drop-shadow(0 0 10px rgb(255 186 79 / 40%)) drop-shadow(0 0 20px rgb(255 140 54 / 26%)) drop-shadow(0 0 36px rgb(142 88 255 / 22%));
   animation: lion-neon-pulse 2.8s ease-in-out infinite;
 }
 
 @keyframes lion-neon-pulse {
-
   0%,
   100% {
-    filter:
-      drop-shadow(0 0 8px rgb(255 186 79 / 34%)) drop-shadow(0 0 16px rgb(255 140 54 / 22%)) drop-shadow(0 0 30px rgb(142 88 255 / 18%));
+    filter: drop-shadow(0 0 8px rgb(255 186 79 / 34%)) drop-shadow(0 0 16px rgb(255 140 54 / 22%)) drop-shadow(0 0 30px rgb(142 88 255 / 18%));
   }
 
   50% {
-    filter:
-      drop-shadow(0 0 12px rgb(255 196 95 / 52%)) drop-shadow(0 0 26px rgb(255 153 66 / 34%)) drop-shadow(0 0 42px rgb(154 100 255 / 30%));
+    filter: drop-shadow(0 0 12px rgb(255 196 95 / 52%)) drop-shadow(0 0 26px rgb(255 153 66 / 34%)) drop-shadow(0 0 42px rgb(154 100 255 / 30%));
   }
 }
 
@@ -154,10 +146,7 @@
   padding: 28px;
   color: #e7efff;
   text-align: left;
-  box-shadow:
-    inset 0 0 26px rgb(67 85 180 / 10%),
-    0 8px 22px rgb(0 0 0 / 38%),
-    0 3px 8px rgb(0 0 0 / 22%);
+  box-shadow: inset 0 0 26px rgb(67 85 180 / 10%), 0 8px 22px rgb(0 0 0 / 38%), 0 3px 8px rgb(0 0 0 / 22%);
   background: linear-gradient(180deg, rgb(8 12 26 / 90%) 0%, rgb(7 10 20 / 92%) 100%);
   border: 1px solid rgb(145 93 255 / 58%);
   position: relative;
